+++ atst/utils/json.py
@@ -1,6 +1,9 @@
 from flask.json import JSONEncoder
+import json
 from werkzeug.datastructures import FileStorage
 from datetime import date
+from enum import Enum
+
 from atst.models.attachment import Attachment
 
 
@@ -13,3 +16,13 @@ class CustomJSONEncoder(JSONEncoder):
         elif isinstance(obj, FileStorage):
             return obj.filename
         return JSONEncoder.default(self, obj)
+
+
+def sqlalchemy_dumps(dct):
+    def _default(obj):
+        if isinstance(obj, Enum):
+            return obj.name
+        else:
+            raise TypeError()
+
+    return json.dumps(dct, default=_default)
